+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## 1.0.6 - 2017-06-20
 
-* New states were added for the management of the modules: `updated` and `outdated`.
+* Bugs fixed in doAction() method and when using the Hook class.
 
 * Deleted `Eliasis\Module\Module->_setState()` method.